admin: extract tests for Pôle Emploi
