+++ mon-pix/tests/acceptance/start-campaigns-workflow_test.js
@@ -11,8 +11,6 @@ import { Response } from 'ember-cli-mirage';
 import { contains } from '../helpers/contains';
 import { clickByLabel } from '../helpers/click-by-label';
 import { fillInByLabel } from '../helpers/fill-in-by-label';
-import sinon from 'sinon';
-import Service from '@ember/service';
 
 import { authenticateByEmail, authenticateByGAR } from '../helpers/authentication';
 import { startCampaignByCode, startCampaignByCodeAndExternalId } from '../helpers/campaign';
@@ -368,96 +366,6 @@ describe('Acceptance | Campaigns | Start Campaigns workflow', function () {
           });
         });
 
-        context('When campaign belongs to Pole Emploi organization', function () {
-          let replaceLocationStub;
-
-          beforeEach(function () {
-            replaceLocationStub = sinon.stub().resolves();
-            this.owner.register(
-              'service:location',
-              Service.extend({
-                replace: replaceLocationStub,
-              })
-            );
-            campaign = server.create('campaign', { identityProvider: identityProviders.POLE_EMPLOI });
-          });
-
-          it('should redirect to landing page', async function () {
-            // given
-            await visit('/campagnes');
-
-            // when
-            await fillIn('#campaign-code', campaign.code);
-            await clickByLabel(t('pages.fill-in-campaign-code.start'));
-
-            // then
-            expect(currentURL()).to.equal(`/campagnes/${campaign.code}/presentation`);
-          });
-
-          it('should redirect to Pole Emploi authentication form when landing page has been seen', async function () {
-            // given
-            await visit(`/campagnes/${campaign.code}`);
-
-            // when
-            await clickByLabel('Je commence');
-
-            // then
-            sinon.assert.called(replaceLocationStub);
-            expect(currentURL()).to.equal('/connexion-pole-emploi');
-          });
-
-          it('should redirect to terms of service Pole Emploi page', async function () {
-            // given
-            const state = 'state';
-            const session = currentSession();
-            session.set('data.state', state);
-            this.server.post('pole-emploi/token', () => {
-              return new Response(
-                401,
-                {},
-                {
-                  errors: [
-                    {
-                      code: 'SHOULD_VALIDATE_CGU',
-                      meta: {
-                        authenticationKey: 'key',
-                      },
-                    },
-                  ],
-                }
-              );
-            });
-
-            // when
-            await visit(`/connexion-pole-emploi?code=test&state=${state}`);
-
-            // then
-            expect(currentURL()).to.equal(`/cgu-pole-emploi?authenticationKey=key`);
-            expect(find('.terms-of-service-form__conditions').textContent).to.contains(
-              "J'accepte les conditions d'utilisation et la politique de confidentialité de Pix"
-            );
-          });
-
-          it('should begin campaign participation once user has accepted terms of service', async function () {
-            // given
-            const state = 'state';
-            const session = currentSession();
-            session.set('data.state', state);
-            session.set('data.nextURL', `/campagnes/${campaign.code}/acces`);
-            const data = {};
-            data[campaign.code] = { landingPageShown: true };
-            sessionStorage.setItem('campaigns', JSON.stringify(data));
-
-            // when
-            await visit(`/cgu-pole-emploi?authenticationKey=key`);
-            await clickByLabel("J'accepte les conditions d'utilisation et la politique de confidentialité de Pix");
-            await clickByLabel('Je continue');
-
-            // then
-            expect(currentURL()).to.equal(`/campagnes/${campaign.code}/evaluation/didacticiel`);
-          });
-        });
-
         context('When is a simplified access campaign', function () {
           beforeEach(function () {
             campaign = server.create('campaign', { isSimplifiedAccess: true, idPixLabel: 'Les anonymes' });
@@ -732,99 +640,6 @@ describe('Acceptance | Campaigns | Start Campaigns workflow', function () {
 
           // then
           expect(currentURL()).to.equal(`/campagnes/${campaign.code}/evaluation/didacticiel`);
-        });
-      });
-
-      context('When campaign belongs to Pole Emploi organization', function () {
-        let replaceLocationStub;
-
-        beforeEach(function () {
-          replaceLocationStub = sinon.stub().resolves();
-          this.owner.register(
-            'service:location',
-            Service.extend({
-              replace: replaceLocationStub,
-            })
-          );
-          campaign = server.create('campaign', { identityProvider: identityProviders.POLE_EMPLOI });
-        });
-
-        context('When user is logged in with Pole emploi', function () {
-          beforeEach(function () {
-            const session = currentSession();
-            session.set('data.authenticated.identityProvider', 'POLE_EMPLOI');
-          });
-
-          it('should redirect to landing page', async function () {
-            // given
-            await visit('/campagnes');
-
-            // when
-            await fillIn('#campaign-code', campaign.code);
-            await clickByLabel(t('pages.fill-in-campaign-code.start'));
-
-            // then
-            expect(currentURL()).to.equal(`/campagnes/${campaign.code}/presentation`);
-          });
-
-          it('should begin campaign participation', async function () {
-            // given
-            await visit('/campagnes');
-            await fillIn('#campaign-code', campaign.code);
-            await clickByLabel(t('pages.fill-in-campaign-code.start'));
-
-            // when
-            await clickByLabel('Je commence');
-
-            // then
-            expect(currentURL()).to.equal(`/campagnes/${campaign.code}/evaluation/didacticiel`);
-          });
-        });
-
-        context('When user is logged in with another authentication method', function () {
-          it('should redirect to landing page', async function () {
-            // given
-            await visit('/campagnes');
-
-            // when
-            await fillIn('#campaign-code', campaign.code);
-            await clickByLabel(t('pages.fill-in-campaign-code.start'));
-
-            // then
-            expect(currentURL()).to.equal(`/campagnes/${campaign.code}/presentation`);
-          });
-
-          it('should redirect to Pole Emploi authentication form when landing page has been seen', async function () {
-            // given
-            await visit(`/campagnes/${campaign.code}`);
-
-            // when
-            await clickByLabel('Je commence');
-
-            // then
-            sinon.assert.called(replaceLocationStub);
-            expect(currentURL()).to.equal('/connexion-pole-emploi');
-          });
-
-          it('should begin campaign participation once user is authenticated', async function () {
-            // given
-            const state = 'state';
-
-            const session = currentSession();
-            session.set('isAuthenticated', true);
-            session.set('data.state', state);
-            session.set('data.nextURL', `/campagnes/${campaign.code}/acces`);
-            const data = {};
-            data[campaign.code] = { landingPageShown: true };
-            sessionStorage.setItem('campaigns', JSON.stringify(data));
-
-            // when
-            await visit(`/connexion-pole-emploi?code=test&state=${state}`);
-
-            // then
-            sinon.assert.notCalled(replaceLocationStub);
-            expect(currentURL()).to.equal(`/campagnes/${campaign.code}/evaluation/didacticiel`);
-          });
         });
       });
